refactor: Slackファイル分離
